requirements.txt
textual-dev
rich-pixels
